first step for migration to typescript
