+++ .gitignore
@@ -1,3 +1,5 @@
+bin
+
 ### Node template
 # Logs
 logs
