+++ package.json
@@ -25,15 +25,18 @@
     "find-up": "^4.1.0"
   },
   "devDependencies": {
+    "@types/node": "^13.1.6",
     "ava": "^2.4.0",
-    "eslint": "^6.7.2",
+    "eslint": "^6.8.0",
     "eslint-config-xo-space": "^0.22.0",
-    "rimraf": "^3.0.0"
+    "rimraf": "^3.0.0",
+    "typescript": "^3.7.4"
   },
   "engines": {
-    "node": ">=8"
+    "node": ">=10.13"
   },
   "scripts": {
+    "compile": "tsc",
     "lint": "eslint \"bin/*.js\"",
     "prepare:husky:2": "cd ./test/husky2 && git init && git checkout -b JIRA-1234-test-husky2 && npm i",
     "prepare:husky:3": "cd ./test/husky3 && git init && git checkout -b JIRA-4321-test-husky3 && npm i",
